tools/requirements.txt
pandas>=2.0.0
neologdn>=0.5.0
supabase>=2.0.0
python-dotenv>=1.0.0
